Complete resturcture. Moving to POSIX.
The descision was made to make this library as POSIX compliant as
possible. All non-POSIX functions are now removed and will probably end
up in some other repository. In the mean time the main focus of this
repository should be to provide sane libc experience using only /globaly/
allowed functions as defined by 42 libft subject file.